requirements.txt
Flask
flask-cors
requests
Pillow
google-generativeai
python-dotenv
gunicorn